+++ benchmark/cmd/main.go
@@ -38,7 +38,6 @@ func main() {
 		cfg.CacheDir = path.SystemCacheDir("sensu-agent")
 		cfg.Deregister = true
 		cfg.DeregistrationHandler = ""
-		cfg.ExtendedAttributes = []byte{}
 		cfg.KeepaliveInterval = agent.DefaultKeepaliveInterval
 		cfg.KeepaliveTimeout = types.DefaultKeepaliveTimeout
 		cfg.Namespace = agent.DefaultNamespace