+++ types/check.go
@@ -47,13 +47,17 @@ var OutputMetricFormats = []string{NagiosOutputMetricFormat, GraphiteOutputMetri
 // and encoding/json.
 func NewCheck(c *CheckConfig) *Check {
 	check := &Check{
+		ObjectMeta: ObjectMeta{
+			Name:        c.Name,
+			Namespace:   c.Namespace,
+			Labels:      c.Labels,
+			Annotations: c.Annotations,
+		},
 		Command:              c.Command,
 		Handlers:             c.Handlers,
 		HighFlapThreshold:    c.HighFlapThreshold,
 		Interval:             c.Interval,
 		LowFlapThreshold:     c.LowFlapThreshold,
-		Name:                 c.Name,
-		Namespace:            c.Namespace,
 		Publish:              c.Publish,
 		RuntimeAssets:        c.RuntimeAssets,
 		Subscriptions:        c.Subscriptions,
@@ -70,12 +74,6 @@ func NewCheck(c *CheckConfig) *Check {
 		OutputMetricHandlers: c.OutputMetricHandlers,
 		EnvVars:              c.EnvVars,
 	}
-	// Unmarshal extended attributes into a different Check value, so that
-	// we don't accidentally corrupt any of the default values for Check.
-	// See https://github.com/sensu/sensu-go/issues/1732 for more information.
-	tmpCheck := Check{}
-	_ = dynamic.Unmarshal(c.ExtendedAttributes, &tmpCheck)
-	check.ExtendedAttributes = tmpCheck.ExtendedAttributes
 	return check
 }
 
@@ -139,11 +137,6 @@ func (c *Check) Validate() error {
 	return c.Subdue.Validate()
 }
 
-// UnmarshalJSON implements the json.Unmarshaler interface.
-func (c *Check) UnmarshalJSON(b []byte) error {
-	return dynamic.Unmarshal(b, c)
-}
-
 // MarshalJSON implements the json.Marshaler interface.
 func (c *Check) MarshalJSON() ([]byte, error) {
 	if c == nil {
@@ -156,12 +149,6 @@ func (c *Check) MarshalJSON() ([]byte, error) {
 		c.Handlers = []string{}
 	}
 
-	// Only use dynamic marshaling if there are dynamic attributes.
-	// Otherwise, use default json marshaling.
-	if len(c.ExtendedAttributes) > 0 {
-		return dynamic.Marshal(c)
-	}
-
 	type Clone Check
 	clone := &Clone{}
 	*clone = Clone(*c)
@@ -169,19 +156,13 @@ func (c *Check) MarshalJSON() ([]byte, error) {
 	return jsoniter.Marshal(clone)
 }
 
-// SetExtendedAttributes sets the serialized ExtendedAttributes of c.
-func (c *Check) SetExtendedAttributes(e []byte) {
-	c.ExtendedAttributes = e
-}
-
 // Get implements govaluate.Parameters
 func (c *Check) Get(name string) (interface{}, error) {
 	return dynamic.GetField(c, name)
 }
 
-// UnmarshalJSON implements the json.Unmarshaler interface.
-func (c *CheckConfig) UnmarshalJSON(b []byte) error {
-	return dynamic.Unmarshal(b, c)
+func (c *CheckConfig) Get(name string) (interface{}, error) {
+	return dynamic.GetField(c, name)
 }
 
 // MarshalJSON implements the json.Marshaler interface.
@@ -196,27 +177,11 @@ func (c *CheckConfig) MarshalJSON() ([]byte, error) {
 		c.Handlers = []string{}
 	}
 
-	// Only use dynamic marshaling if there are dynamic attributes.
-	// Otherwise, use default json marshaling.
-	if len(c.ExtendedAttributes) > 0 {
-		return dynamic.Marshal(c)
-	}
-
 	type Clone CheckConfig
 	clone := &Clone{}
 	*clone = Clone(*c)
 
 	return jsoniter.Marshal(clone)
-}
-
-// SetExtendedAttributes sets the serialized ExtendedAttributes of c.
-func (c *CheckConfig) SetExtendedAttributes(e []byte) {
-	c.ExtendedAttributes = e
-}
-
-// Get implements govaluate.Parameters
-func (c *CheckConfig) Get(name string) (interface{}, error) {
-	return dynamic.GetField(c, name)
 }
 
 // Validate returns an error if the check does not pass validation tests.
@@ -359,15 +324,17 @@ func FixtureCheckConfig(id string) *CheckConfig {
 	interval := uint32(60)
 	timeout := uint32(0)
 
-	return &CheckConfig{
-		Name:                 id,
+	check := &CheckConfig{
+		ObjectMeta: ObjectMeta{
+			Name:      id,
+			Namespace: "default",
+		},
 		Interval:             interval,
 		Subscriptions:        []string{"linux"},
 		Command:              "command",
 		Handlers:             []string{},
 		RuntimeAssets:        []string{"ruby-2-4-2"},
 		CheckHooks:           []HookList{*FixtureHookList("hook1")},
-		Namespace:            "default",
 		Publish:              true,
 		Cron:                 "",
 		Ttl:                  0,
@@ -375,6 +342,7 @@ func FixtureCheckConfig(id string) *CheckConfig {
 		OutputMetricHandlers: []string{},
 		OutputMetricFormat:   "",
 	}
+	return check
 }
 
 // FixtureCheck returns a fixture for a Check object.