+++ types/dynamic/redact.go
@@ -1,8 +1,6 @@
 package dynamic
 
 import (
-	"encoding/json"
-	"fmt"
 	"reflect"
 
 	utilstrings "github.com/sensu/sensu-go/util/strings"
@@ -17,31 +15,25 @@ var DefaultRedactFields = []string{"password", "passwd", "pass", "api_key",
 
 // Redact recursively loops through v in order to redact sensitive fields
 // specified in fields and returns the redacted version of v
-func Redact(v Attributes, fields ...string) (Attributes, error) {
+func Redact(v interface{}, fields ...string) (interface{}, error) {
 	original := reflect.Indirect(reflect.ValueOf(v))
 	redacted := reflect.New(original.Type()).Elem()
-
-	// Retrieve the address of the extended attributes so we can deal with them
-	// appropriately
-	attrsAddress := addressOfExtendedAttributes(v)
 
 	// Use the default fields to redact in case we received none
 	if len(fields) == 0 {
 		fields = DefaultRedactFields
 	}
 
-	if err := redactValue(original, redacted, "", fields, attrsAddress); err != nil {
+	if err := redactValue(original, redacted, "", fields); err != nil {
 		return nil, err
 	}
 
-	redactedAttribute, _ := redacted.Addr().Interface().(Attributes)
-
-	return redactedAttribute, nil
+	return redacted.Addr().Interface(), nil
 }
 
 // redactValue is a recursive function that redacts the original value, into
 // redacted value, based on its type
-func redactValue(original, redacted reflect.Value, fieldName string, fields []string, attrsAddress *byte) error {
+func redactValue(original, redacted reflect.Value, fieldName string, fields []string) error {
 	// Verify if this field is configured to be redacted
 	if utilstrings.FoundInArray(fieldName, fields) {
 		if original.Kind() == reflect.Interface {
@@ -61,50 +53,21 @@ func redactValue(original, redacted reflect.Value, fieldName string, fields []st
 		return nil
 	}
 
-	// Verify if we are dealing with the extended attributes
-	if !isEmpty(original) && isExtendedAttributes(attrsAddress, original) {
-		// Unmarshal the extended attributes so we can deal with them
-		var attrs interface{}
-		if err := json.Unmarshal(original.Bytes(), &attrs); err != nil {
-			return fmt.Errorf("could not unmarshal the redacted version of extended attribute: %s", err)
-		}
-
-		// Get the underlying value of the extended attributes
-		originalValue := reflect.ValueOf(attrs)
-
-		// Initialize a new object with the extended attributes type
-		redactedValue := reflect.New(originalValue.Type()).Elem()
-
-		// Redact values within the extended attributes
-		if err := redactValue(originalValue, redactedValue, fieldName, fields, attrsAddress); err != nil {
-			return err
-		}
-
-		// Marshal back the extended attributes so we can set them back into redacted
-		bytes, err := json.Marshal(redactedValue.Interface())
-		if err != nil {
-			return fmt.Errorf("could not marshal the redacted version of extended attribute: %s", err)
-		}
-
-		redacted.SetBytes(bytes)
-		return nil
-	}
-
 	switch original.Kind() {
 	case reflect.Interface:
-		if err := redactInterface(original, redacted, fieldName, fields, attrsAddress); err != nil {
+		if err := redactInterface(original, redacted, fieldName, fields); err != nil {
 			return err
 		}
 	case reflect.Map:
-		if err := redactMap(original, redacted, fieldName, fields, attrsAddress); err != nil {
+		if err := redactMap(original, redacted, fieldName, fields); err != nil {
 			return err
 		}
 	case reflect.Ptr:
-		if err := redactPtr(original, redacted, fieldName, fields, attrsAddress); err != nil {
+		if err := redactPtr(original, redacted, fieldName, fields); err != nil {
 			return err
 		}
 	case reflect.Struct:
-		if err := redactStruct(original, redacted, fieldName, fields, attrsAddress); err != nil {
+		if err := redactStruct(original, redacted, fieldName, fields); err != nil {
 			return err
 		}
 	default:
@@ -117,7 +80,7 @@ func redactValue(original, redacted reflect.Value, fieldName string, fields []st
 
 // redactInterface retrieves the actual value and type of original interface and
 // calls back redactValue with the right value
-func redactInterface(original, redacted reflect.Value, fieldName string, fields []string, attrsAddress *byte) error {
+func redactInterface(original, redacted reflect.Value, fieldName string, fields []string) error {
 	// Get actual value of original by getting rid of the interface
 	originalValue := original.Elem()
 
@@ -130,7 +93,7 @@ func redactInterface(original, redacted reflect.Value, fieldName string, fields 
 	redactedValue := reflect.New(originalValue.Type()).Elem()
 
 	// Redact values within the interface value
-	if err := redactValue(originalValue, redactedValue, fieldName, fields, attrsAddress); err != nil {
+	if err := redactValue(originalValue, redactedValue, fieldName, fields); err != nil {
 		return err
 	}
 
@@ -141,7 +104,7 @@ func redactInterface(original, redacted reflect.Value, fieldName string, fields 
 
 // redactMap loops through every key of the original value map in order to
 // redact any of these keys found in fields
-func redactMap(original, redacted reflect.Value, fieldName string, fields []string, attrsAddress *byte) error {
+func redactMap(original, redacted reflect.Value, fieldName string, fields []string) error {
 	// Make sure it's not empty
 	if isEmpty(original) {
 		return nil
@@ -159,7 +122,7 @@ func redactMap(original, redacted reflect.Value, fieldName string, fields []stri
 		redactedValue := reflect.New(originalValue.Type()).Elem()
 
 		// Redact values in that key
-		if err := redactValue(originalValue, redactedValue, key.String(), fields, attrsAddress); err != nil {
+		if err := redactValue(originalValue, redactedValue, key.String(), fields); err != nil {
 			return err
 		}
 
@@ -169,7 +132,7 @@ func redactMap(original, redacted reflect.Value, fieldName string, fields []stri
 	return nil
 }
 
-func redactPtr(original, redacted reflect.Value, fieldName string, fields []string, attrsAddress *byte) error {
+func redactPtr(original, redacted reflect.Value, fieldName string, fields []string) error {
 	// Get actual value of original by getting rid of the pointer
 	originalValue := original.Elem()
 
@@ -182,12 +145,12 @@ func redactPtr(original, redacted reflect.Value, fieldName string, fields []stri
 	redacted.Set(reflect.New(originalValue.Type()))
 
 	// Redact values within the pointer value
-	return redactValue(originalValue, redacted.Elem(), fieldName, fields, attrsAddress)
+	return redactValue(originalValue, redacted.Elem(), fieldName, fields)
 }
 
 // redactStruct loops through every field of the original value struct in order
 // to redact any of these fields found in fields
-func redactStruct(original, redacted reflect.Value, fieldName string, fields []string, attrsAddress *byte) error {
+func redactStruct(original, redacted reflect.Value, fieldName string, fields []string) error {
 	// Make sure it's not empty
 	if isEmpty(original) {
 		return nil
@@ -195,7 +158,7 @@ func redactStruct(original, redacted reflect.Value, fieldName string, fields []s
 
 	// Loop through every field and call back redactValue
 	for i := 0; i < original.NumField(); i++ {
-		if err := redactValue(original.Field(i), redacted.Field(i), original.Type().Field(i).Name, fields, attrsAddress); err != nil {
+		if err := redactValue(original.Field(i), redacted.Field(i), original.Type().Field(i).Name, fields); err != nil {
 			return err
 		}
 	}