+++ types/entity_test.go
@@ -36,15 +36,9 @@ func TestFixtureEntityIsValid(t *testing.T) {
 
 func TestEntityGet(t *testing.T) {
 	e := FixtureEntity("myAgent")
-	e.SetExtendedAttributes([]byte(`{"foo": "bar"}`))
-
-	// Find extended attr
-	val, err := e.Get("foo")
-	require.NoError(t, err)
-	assert.EqualValues(t, "bar", val)
 
 	// Find exported field
-	val, err = e.Get("ID")
+	val, err := e.Get("ID")
 	require.NoError(t, err)
 	assert.EqualValues(t, "myAgent", val)
 }
@@ -53,27 +47,19 @@ func TestEntityUnmarshal(t *testing.T) {
 	entity := Entity{}
 
 	// Unmarshal
-	err := json.Unmarshal([]byte(`{"id": "myAgent", "foo": "bar"}`), &entity)
+	err := json.Unmarshal([]byte(`{"id": "myAgent"}`), &entity)
 	require.NoError(t, err)
 
 	// Existing exported fields were properly set
 	assert.Equal(t, "myAgent", entity.ID)
-
-	// ExtendedAttribute
-	f, err := entity.Get("foo")
-	require.NoError(t, err)
-	assert.EqualValues(t, "bar", f)
 }
 
 func TestEntityMarshal(t *testing.T) {
 	entity := FixtureEntity("myAgent")
-	entity.SetExtendedAttributes([]byte(`{"foo": "bar"}`))
 
 	bytes, err := json.Marshal(entity)
 	require.NoError(t, err)
 	assert.Contains(t, string(bytes), "myAgent")
-	assert.Contains(t, string(bytes), "foo")
-	assert.Contains(t, string(bytes), "bar")
 }
 
 func TestSortEntitiesByID(t *testing.T) {